Initial Commit!
Going to try out Rust again. This time, after looking into C and C++ again, I think I will like Rust more, even with it's trade-offs that are initially nudging me towards using C, since it allows for more leeway in terms of coding styles and opinions. I think you get a lot more maintainability with Rust though; for me Cargo is the main beneficial thing, as well as the borrow checker. I'm not sure how I feel about snake_case or indentation and things like that, but much cooler expandable types and such are extra intriguing to me.

Yeah so I want to try implementing an NBT implementation in Rust, this case it's more for the sake of seeing how it might work in Rust, as well as getting me to learn some Rust in general.

https://github.com/zugebot/LegacyEditor
https://stackoverflow.com/questions/1353022/reflection-support-in-c (I'd like to be able to log objects to the console, doesn't seems like C-like has that?)
https://www.reddit.com/r/rust/comments/bya8k6/programming_with_rust_vs_c_c/
https://doc.rust-lang.org/rust-by-example/custom_types.html (This is what inspired me to want to look into Rust again, I'm curious how objects can work here, and how using Rust primitives can work for NBT's building blocks. I want to bring NBTify's representation model to native code, and see whether it works or not)
https://medium.com/dwelo-r-d/using-c-libraries-in-rust-13961948c72a (https://www.google.com/search?q=use+c+libraries+in+rust, can you do this? I'd like to include zlib (Rust probably already has that?), and some sort of lzx and leveldb implementations as well.)
https://rust-lang.github.io/async-book/ (Didn't dive into this section yet, but I would like to see how Rust handles async code; I have really started to like how JS/TS handles it)
https://github.com/github/gitignore/blob/main/Rust.gitignore (For this project; `cargo new rust_nbt`, then I added this gitignore template)
https://www.reddit.com/r/rust/comments/hhq5ar/even_better_toml_proper_toml_support_for_visual/ (Yay!)
https://marketplace.visualstudio.com/items?itemName=tamasfe.even-better-toml