Module Namespaces + Real-Time Linting
Unfortunately it looks like Rust Analyzer doesn't support real-time linting like TypeScript and others do, which is unfortunate. You have to save the file for the changes to appear. I've also noticed before that you don't get linting support when editing an arbitrary `.rs` file outside of a project folder, which is unfortunate too. I like that with TS, you can do Ctrl+N, and demo out something to see what happens without even needing to save the file, nor have a `package.json` or `tsconfig.json` adjacent to it, for the editor linting to work.

https://stackoverflow.com/questions/65800478/how-to-enable-real-time-linting-while-typing-with-rust-analyzer-in-vscode
https://github.com/rust-lang/rust-analyzer/issues/4185