rustfmt
rustfmt ./src/*

Okay, I'm gonna see if I can make myself get on board with the default formatting, that seems like that would be the responsible thing to do. I tend to want to do things my own way a lot of the time. I want to learn this outright though!!