Enum Distinctions
Looking into whether to use plain primitives for the types of tags, or an enum derived of all the values themselves. GPT seems to recommend the enum route because it allows for more composability and extensibility down the road. The primitives layout seems like a great way to me too though, because you can do simple math with them, they have less overhead, yeah.

https://dev.to/ghost/rust-project-structure-example-step-by-step-3ee
https://stackoverflow.com/questions/46388386/what-exactly-does-derivedebug-mean-in-rust