Marker Types
This suggestion may be the solution to this issue, but I don't quite see if it will be successful or not yet. I think I want to leave this one aside for the time being. This same kind of issue is present over in TS as well, just for CompoundTag types. It's like generics get funky in languages, when you want to constrain them to certain types.

https://users.rust-lang.org/t/binding-struct-generic-type-parameter-by-enum/84639/4
https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=4b287dcbd2034777fd220fa2e58146a3
https://github.com/Offroaders123/NBTify/issues/19