Unused Code + Package Root 2
So the concept here is actually awesome! So the code is unused, regardless of whether it's public to the *library or not. It is only checking if you yourself are using these values. So this is great! You can get around this by either providing a unit test setup to ensure that the implementation is working correctly, or provide a comment block to document the feature's usage. That will also be tested with the `cargo test` command, in and of itself is also a great thing.

https://stackoverflow.com/questions/26946646/package-with-both-a-library-and-a-binary
https://stackoverflow.com/questions/57756927/rust-modules-confusion-when-there-is-main-rs-and-lib-rs