Immutable Vector Initialization
I wasn't sure in the previous commit how to make a vec with content without making it mutable. This makes more sense! Things are clicking.

The second CD to Empath kind of sounds like what PowerNerd ended up being, and just wow, is it clicking!!! Total Collapse is outstanding. Context for what things are really help with understanding it. This album was in a different time for Dev but it feels like it comes with the same honesty that PowerNerd has.

https://stackoverflow.com/questions/71165771/add-element-to-immutable-vector-rust
https://stackoverflow.com/questions/41366896/how-to-make-a-rust-mutable-reference-immutable
https://www.reddit.com/r/rust/comments/fs1g5h/vec_or_vecnew/