FS File to Memory Vector
Now this can read NBT files from a set of binary data, rather than needing the file system itself to operate on top of!

https://doc.rust-lang.org/std/io/struct.Cursor.html

This is the kind of thing I want from libraries to be able to build with. In a way, it's another form of dependency injection I feel, because where the data comes from shouldn't change the way the program runs.

The Way Home!, Numbered, Awake!!, No Twilight Within The Courts Of The Sun