Import Flattenings
Not sure if this is the way people do things, but coming from JS/TS, I like the notion of explicit imports, and not using namespaces as much, since it makes it more clear what exact features you're using, and what you're not :)

Their Colors Fade: Calculator onwards, You Must Be This Tall

The last run of songs today has been kind of a revisit to the last year or two of music discoveries, around springtime I'd say.

I really admire Keneally for his absolute outstanding ability to write and perform synchronized runs on the different instruments he plays. I haven't been able to achieve that myself yet, because I always end up changing the counterpart to something that blends with the first one, rather than a plain mirror of it.