Flattened Library Exports
This makes the root of the library appear flattened like it does in TS and JS. This seems to be the pattern I am used to. I'm not sure whether I will be using this for the exports, or whether I will stick to namespacing with the module names.

The abilities with Rust's module selection appears to be nearly exactly what I was hoping to do with NBTify in TS, this has more ergonomics in doing that.

hey mom, look! im learning a new language!!!