Module Organizing
It doesn't seem like Rust has the same rules of defining things in a certain order like JS does. Say for example, you can reference a class before it is defined. I'm making the educated guess it's because it does this arrangement for you, because things aren't run as they are interpreted, so you can ensure that things are defined in the proper order that they are referenced? Really cool. That's one side-effect of JS I do like actually, it ensures you know when certain parts of your code are referenced across each other, so you organize it in such a way that it is referenced in the order that it gets called in.