Initial Compression Setup
Having to add a compression implementation/abstraction here too. This was also helped from ChatGPT.

I am not using AI to just get it working, I also want to see how it would be done, and do it like that instead. Because when I try to do things with the limited knowledge I do have (we all have), then I try to do it in a way that doesn't work, then I get frustrated and want to stop working on it. I want to have something that is working that does what I want it to, and being able to learn why it works, rather than building a tower from nothing.