IndexMap Object Ordering
Ended up moving over to IndexMap from HashMap, since this preserves the insert order for keys in the objects. This (for the meantime at least) will help ensure that the NBT structures are read and written in the exact same order. I really like to validate that files are written symmetrically to the way they were originally written.

https://stackoverflow.com/questions/45894401/are-there-any-hashmap-implementations-with-consistent-ordering-between-program-r
https://github.com/indexmap-rs/indexmap

Degradation Trip
Hellbound (interesting, kind of sounds like Rainier Fog in some ways!)