Total Formatting
Okay, sticking with the full formatting from the defaults.

Not sure how I want to deal with async code yet, or if I really need it in this context. If I hope to use this in JS later on by way of WASM, I'm not sure how that will work with async there too.