Diff Demo
This does appear to be reading and writing things in the same order, and the assertion does check it is valid! Great! This is just because I wasn't sure it was properly checking the two values, I intentionally made them different to see if it would throw an error, and it does.

Thrilled to starting building things with this, I already want to make some kind of npm package on top of Rust, really excited for that. Same with Tauri too, oh man that opens a-whole-nother door of things to try! Man that will be fun. I wonder if there is any way to run Tauri with WASM, I mean the backend of it. Not sure that really makes any sense, I don't think it does, since it's meant to be the actual native backend of the app itself.