Compression Test Module
Added a test to verify that everything is being compressed and decompressed accordingly.

Looks like using `[..]` just ensures that these iterable-like types are the same all of the time, so you don't have to worry if one is a buffer, one is a slice, one is a view, one is an iterator, etc. It is just easier to always use it then to worry about whether they are the same type.

https://doc.rust-lang.org/rust-by-example/testing/unit_testing.html