NBTify Function Symmetricality
Moved the implementation names over to match that of NBTify, since this is essentially intended now to be a native port of NBTify to Rust. I didn't initially plan for this to be the case, but conceptually I want to try it out because I think it will help me learn the differences between Rust and TS, as that is my main set of knowledge for programming up to this point. While you can't always prevent it from occurring, I want to not have many misconceptions on how I think things work.