TagID Enum
Now the IDs/types for each of the tags are explicit numbers, like in NBTify. This makes things more safe from a usage standpoint, and it allows for an easier usage with match statements and control flow. The usage for error handling is making me rethink how it should be structured in the parsers too, like in what should be an error and what should be a value. Since Rust doesn't have `null`, you instead return an error, and that's what gets passed along instead. So the errors in this case happen earlier down the chain. I think I can still have my intended error usability messages by using `.unwrap()` and methods like that, or rather the ones that customize the error message, I don't mean to remove it. I don't remember the method name for that yet though.

I'd Google it but I'm at the dmv hehe