TryFrom Trait
Not sure whether this or the previous is the way I should do it, or if it's opinionated. Most of Rust isn't opinionated, so I just want to find out what is commonly done. I think this feels more accurate to what I've seen of the language so far, don't know for sure though.