Byte Arrays Wrapper Type
This is because they should be separately distinguished from the plain Vec<tag> usage, since that would just be a ListTag in the end.

Coming back to this project to learn how I will want to implement this for my Chainmail project.

It's not just a goal of solely porting NBTify itself to Rust, but also just making a better package as a whole. I want to try to use an existing Rust NBT package, but I'm not sure it will provide the tooling that I need to be able to build around it with the strong types that I need to define.