TagId Casing (Delve into Remasters)
Just slowly getting my mind over to thinking in Rust syntax conventions, still a bit stubborn in that regard.

Been listening to Selling England By The Pound. Great album.

Really starting to see the difference in intent, and possible opinions between original releases, compared to remasters. It kind of bums me out the weight to which the gravity of it really determines how much it influences the sound. I really don't seem to enjoy compressed work, now that I can see and articulate that that's the issue. It explains why a bunch of albums used to sound good to me, but "they no longer do anymore". It's because I am listening to the remaster, and I didn't know what that meant.

As all things, there are intricasies. Sometimes the remaster is better than the original, and vice versa. The point is, it is still up to your interpretation. So just like "right or wrong" in a music context, it is your opinion. It is up for you to decide whether you like the sound or not, or what it is what you resonate with.

The Sluggo remix for example, miles better than the original. It had dynamic range still in mind when being made (because that's the nature of Keneally's work). For the 2007 Selling England By The Pound, it does indeed sound "clearer", as the general rule of thumb for remasters tend to show. But it also removes a lot of dynamic range, and hearing the dynamic range of the original, I tend to prefer that. I really do like the frequencies brought up, as well as the stereo widening, but I can't lose dynamic range. It adds so much heart and life which the other additions can't really account for. So for that release, I enjoy the original more. The remaster does sound great, I enjoyed listening to it. But after hearing how the original sounded, it isn't keeping around parts of what made the original, the original (in my opinion, as stated).

So maybe this is the outline I might possibly follow going forward (borrowed from programming in concept): If the remaster is unopinionated, I may like it more than the original. If it is opinionated, I might not. I will put that to the test.

If anything, I know it's still opinionated to release a remaster, because that's an opinion in that the original work needed more to be fixed.

But what I mean in an "opinionated remaster" is that the remaster wouldn't change more than one metric at a time. For example, I'd say maybe it just changes the EQ for example, or just makes it feel more space in the stereo field. Maybe even both, right now I feel that those two are "okay". Unless it was too extreme for example. But for me, changing the dynamics, that seems something fairly heavy handed, if done to a large extent. If it changes it to feel like certain passages are no longer there, because they all have the same loudness, then to me that changes the character of the sound. The same could be said of the EQ and stereo field too though. So, it's a wash in the end.

I'm just glad to have built my ability to discern these things, because now that I can see them, it brings much more to the table for me to learn, as well as to create and articulate.

I want to create a very dynamic record now. I notice that a lot of records are probably compressed because people architect them to be made for "listening anywhere". Why is this the case? Is it because people don't value music to where they sit down and listen to it outright, in a quiet room? If so, that absolutely explains why the mixing/mastering changed in that regard. Maybe an album *doesn't* need to be made for listening anywhere. Maybe that's not the actual goal. Maybe it is meant for listening on specific gear, like a stereo, or headphones. If it doesn't sound good in the car on the loud freeway, then so be it. It might be being "musicked wrong", as Adam Neely has pointed out.

Whole new ballgame just started. Let's see how this game goes.
[metaphor]