matches!() Macro!
It's awesome!

Addicted!

Awake!!

AT NIGHT!!!

NOSFERATU!!!!
🧛‍♂️

But yeah, this is something I've wanted for a while, glad to have stumbled upon it. It's the ability to check enum variants like you can as normal values, say like in TS. You don't need to use match (I know it still does under the hood, but yeah). If you only have a single variant to check, then an if statement always felt the most ergonomic to me. This brings that back. Heck yeah!

I do enjoy the syntax of the plain match though, it does make it really clear what is going on. And it is strongly typed as to handling all cases too. I am curious if this one will still get mad at you if you don't handle the else. I don't think it does, I think it implicitly adds it for you. Maybe that's a case against using this one, that it doesn't make you define an else handler (or additional cases, if you'd rather).