Little Endian Test
Going to separate these two into two separate tests, each of the file types.

Will work on making an abstraction for this once it is more clear how my test suite will be designed.

Do I want to include tests in the source files themselves (with test modules), or in a file like this one?

https://doc.rust-lang.org/book/ch11-03-test-organization.html