Archive: Demo Deserializer
I don't plan to keep this one, I still have to go through the code to understand it. The goal here is to be able to go between a Rust type definition, and my NBT primitives, by way of simply using a derive macro, which allows you to not need to write conversion code yourself. It seems that Serde is perfect for this.

Note that the goal is not to use Serde for NBT file construction itself, instead it is for converting between the data formats which are used to describe the certain known/unknown parts of the serialization process.

I don't know if this code works yet, I got help from GPT after explaining how I want things to function. I had to modify some parts to use my IndexMap usage, since it was recommending BTreeMap.