serde::Serialize Flattening
Now it is just an import.

Going back and loving music I have grown up with. The Madness of Many is my favorite AAL album. Joy of Motion is the one that I grew up with, or rather, found them with. But I think MoM is the definitive one for myself. I got it on vinyl not too long ago too, I think that only solidified it as being my personal favorite.

I'd have to say Joy of Motion is the definitive, in my opinion, for the general scheme of things though. Like if I were to show someone them for the first time, it is JoM.

I love Weightless too. All albums are great. I think my current rating, for myself though, in terms of my own connection to the albums as a whole, being song flow, interest throughout the whole album:

- The Madness of Many
- Joy of Motion
- Weightless
- Animals as Leader [self-titled]
- Parrhesia

I am bummed to put the newest album last. I love the songs, but I am now seeing that I do not love the mastering. It is way to flat for my taste. It loses the grit I have loved the band for.

And this list is also the specfic order I might randomly listen to AAL in, like if I were to listen to them, what would I grab first.

Definitive for sharing with others, this list would be much different. It always depends on the criteria of what the list is sorted by.