Array-like Seq Des Revert
Realizing a different way I plan to/have to implement this for the array-like tags.

Just got Deconstruction on Bandcamp. After about a year of checking out the songs, now I think I understand the "point" of the album. I loved the music to start with, now I get what he is trying to convey, and or communicate. It's something I have been trying to work through myself as well. And maybe I am understanding something out of it that's different than what it's meant to convey. But either way, I have a new understanding of myself, and maybe that's why I am able to hear Deconstruction. Or it could be some other random thing I don't have an answer to. And maybe I don't need an answer? That's kind of the point of my thinking here, not everything you do needs to have a reason, or if anything, maybe you just don't need to know the reason. Not all things have to be meticulously planned, just live your life, spur of the moment!