Possible ByteArrayTag Deserialization
Okay, this might actually be working. We'll have to wait and see until I implement this for the other two typed-array tags though, because it is still erroring at those.

At least I am getting to IntArrayTag now, that's progress.

Getting help from GPT now.

The new story in the news makes me sick about AI though, just as I was talking the other day about using it healthily for healthy things, I see that it really has a large amount of power, and it scares me, as well as it being just daunting and heavily robotic. It is scary what it is, not that it's sentient, but more that it operates very psychopathically, in the nature of how it works. I am so sorry to the young man we lost and his parents and family, it's no laughing matter what AI is capable of nowadays. It's absolutely up to us to know what it's strengths and weaknesses are. It is very much a "yes" machine. I have noticed this too, that it will try to agree with you on everything. It is very much an enabler.