Serde Serializers Revert
This is the revert of the second experimentation for my Serde implementation. I very much still want to get it to work, I think I need to rework the underlying types and implementations before I do that though, as right now I seem to be implementing myself into a corner, with things that don't quite work accordingly.