VarInt + Unnamed Fixtures
Going to omit the SNBT one for until later. That isn't in the scope of things right now. I will come back to it. Of course I do want SNBT support, but with all of the spec changes Mojang made in the last few updates, it's kind of a mess to deal with in terms of compatibility now, which is a real shame. Rather than keeping it a nice format, it does somewhat of a compile step with mappings of certain object shapes. Seems not very deterministic to me, again, bummer. I know mapping-wise it is probably deterministic, but I mean in that what NBT you would have before, now it doesn't always map to the same SNBT anymore. So how do you version NBT files? Yuck, just a whole mess. Maybe I just won't worry about the "old" SNBT behavior? We will see. Will have to think about it some more.