Endian Enum
Working out how to make this the deciding factor to how the reader functions are called, rather than having the byteorder crate be the deciding part of the API. It's the `LittleVarInt` option that makes this confusing. Been trying for like 2 hours to try and see how to make this work.

Sad with how the "end of times" is currently upon us. It's that we are in a new era, and I don't know how it works. People we love, we have to see them leave us. We are now the ones people look up to. How is this the case, and now what? I am hoping for the most love to come of this. Loss is very hard.