VarInt Trait Demo
This was help from GPT. Adding this into the source control for future reference. It seems to work how I was hoping to make it work, I just wasn't sure how to do it, if it was possible, or if my goal of wanting to do it this way was accurate. It seems at least now I know it's possible!