Endian Param
Instead of generics. This allows for specifying VarInt as an option now.

I liked the generic usage, but it's alright. It might not have necessarily been the correct way to make the API public though, so this makes a bit more sense now. It's more strongly-typed, which is probably a good thing.