VarInt Test Demo
Okay, made a file to test the raw encodings against plainly.

From just loading this the first time, I feel like it may just be a simply fix, to some extent. Will see.

Loving Ghost, on Dark Matters/Texada. Really love the sound and feeling of this part here.
The entire DTP trip is almost complete. When the time is right, I will go back and experience Ki. I was going to go sequentially, but I ended up going as I felt connection to the albums, and in relation to the things in my life at the time. I started with Addicted, then attempted Ki and Ghost in parts, didn't stick the first time. Later on, got to Deconstruction. Still settling in with that one. Love it a lot. Totally digging Ghost now, this is the time for it. Can't wait to get it on Bandcamp, listening from YouTube at the moment.

Always have compassion for yourself.