VarInt ZigZag Reimpl
This is the rewrite of the code that wasn't working.

Infinite Ocean is awesome, I love it.

This is not passing yet, it is because of the logic at the end of the function to perform the ZigZag step. Will debug it separate from this. Just wanted to see if moving the one over from JS would work. Looks like not.